refactor: 주문과 주문 내의 아이템으로 도메인 구분
