+++ src/main/java/store/domain/OrderItem.java
@@ -1,0 +1,12 @@
+package store.domain;
+
+public class OrderItem {
+    private final String productName;
+    private final int quantity;
+
+    public OrderItem(final String productName, final int quantity) {
+        this.productName = productName;
+        this.quantity = quantity;
+    }
+
+}
