--- src/main/java/store/domain/Orders.java
@@ -1,11 +1,0 @@
-package store.domain;
-
-import java.util.List;
-
-public class Orders {
-    private final List<Orders> orders;
-
-    public Orders(final List<Orders> orders) {
-        this.orders = orders;
-    }
-}
